Duration type changed to int.

+++ lamp_kotlin/src/main/java/digital/lamp/lamp_kotlin/lamp_core/models/DimensionData.kt
@@ -104,7 +104,7 @@ data class LowPowerModeData(
 )
 
 data class TelephonyData(
-    val duration: Double?,
+    val duration: Int?,
     val trace: String?,
     val type: String?
 )
